Update - JS
My brain hit a brick wall with this one. I need to convert the prompts for number of players and starting life and see if i can figure out how to break them up into smaller methods or functions. That and creating divs = number of players using a for loop,  then trying to add an h1 and span sucks. So I'm going to get some sleep and come back to this issue at a later date